requirements.txt
scikit-image==0.18.1
numpy==1.19.5
matplotlib==3.4.1
Pillow==8.2.0
efficientnet==1.1.1
streamlit==0.81.1
Keras==2.4.3
tensorflow==2.4.1